chore: rename soar.client.py to soar.service.py
